refactor: move get-transaction and get-transaction_test to query to use cqrs pattern :hammer:
