+++ components/transaction/internal/app/query/get-transaction.go
@@ -1,11 +1,11 @@
-package command
+package query
 
 import (
 	"context"
-
 	"github.com/LerianStudio/midaz/common/mlog"
 	t "github.com/LerianStudio/midaz/components/transaction/internal/domain/transaction"
 	"github.com/google/uuid"
+	"reflect"
 )
 
 // GetTransactionByID gets data in the repository.
@@ -17,6 +17,18 @@ func (uc *UseCase) GetTransactionByID(ctx context.Context, organizationID, ledge
 	if err != nil {
 		logger.Errorf("Error getting transaction: %v", err)
 		return nil, err
+	}
+
+	if tran != nil {
+		metadata, err := uc.MetadataRepo.FindByEntity(ctx, reflect.TypeOf(t.Transaction{}).Name(), transactionID)
+		if err != nil {
+			logger.Errorf("Error get metadata on mongodb account: %v", err)
+			return nil, err
+		}
+
+		if metadata != nil {
+			tran.Metadata = metadata.Data
+		}
 	}
 
 	return tran, nil
