+++ components/transaction/internal/app/query/get-transaction_test.go
@@ -1,4 +1,4 @@
-package command
+package query
 
 import (
 	"context"
